Fix a few errors in the readme and stop critising the original dev
The commit should fix minor spelling and grammar errors in the readme + remove all critising comments for the original dev.

I may get thrashed for the second bit, but after I noticed the cause of the feud here:

https://github.com/kauancvlcnt/Zorin-OS-Pro/issues/7

Everything goes down the issue of not being credited.

Ok, its understandable, but such oversights are quite common in open source repos. I myself have been both a victim as well as have unknowingly done the same to other, but we never make a fuss about it. The main goal always remains to get a working software for the community.

Also, at the time of the PR, there was no license involved, so whatever was done was legal.

Anyways the original repo has been archived. So, lets just thank the original dev for whatever he did and forget the bad encounters.

No offence intended to anyone, and obviously I can add back the critising comments if requested.